library/requirements.txt
pandas
psycopg2
python-dotenv
pre-commit
boto3
black
sqlalchemy